Split out components
