+++ src/components/Loading.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Loading = (props) => {
+    return (
+      <div style={{position: 'absolute', left: '50%', top: '50%', transform: 'translate(-50%, -50%)' }}>Loading...</div>
+    )
+  }
+  
+export default Loading;
